Move the source dependency resolution infrastructure out of the dependency management project and into version control project.
