+++ subprojects/language-native/language-native.gradle
@@ -25,6 +25,7 @@ dependencies {
     compile project(':platformNative')
     compile project(':maven')
 
+    implementation project(':versionControl')
     implementation libraries.commons_io.coordinates
 
     integTestRuntimeOnly project(":ideNative")
