+++ subprojects/version-control/src/main/resources/META-INF/services/org.gradle.internal.service.scopes.PluginServiceRegistry
@@ -1,1 +1,1 @@
-org.gradle.vcs.internal.VersionControlServices
+org.gradle.vcs.internal.services.VersionControlServices
