+++ subprojects/version-control/version-control.gradle
@@ -25,6 +25,7 @@ dependencies {
     api project(':coreApi')
 
     implementation project(':core')
+    implementation project(':dependencyManagement')
     implementation libraries.jgit.coordinates, libraries.commons_httpclient.coordinates, libraries.jsch.coordinates
 
     testFixturesImplementation project(":internalTesting")
